评论回复 展示及发表

+++ src/main/java/com/ce/majiang/controller/PublishController.java
@@ -38,7 +38,7 @@ public class PublishController {
             @RequestParam(name = "title", required = false) String title,
             @RequestParam(name = "description", required = false) String description,
             @RequestParam(name = "tag", required = false) String tag,
-            @RequestParam(name = "id", required = false) Integer id,
+            @RequestParam(name = "id", required = false) Long id,
             HttpServletRequest request, Model model) {
         model.addAttribute("title", title);
         model.addAttribute("tag", tag);
